#70761. Some refactoring as per Code Review comments. Phase III.

+++ src/netstandard/RaptorClassLibrary.netstandard/Types/CMVRangePercentageRecord.cs
@@ -55,8 +55,8 @@ namespace VSS.TRex.Types
     /// <param name="reader"></param>
     public void Read(BinaryReader reader)
     {
-      Min = reader.ReadUInt16();
-      Max = reader.ReadUInt16();
+      Min = reader.ReadDouble();
+      Max = reader.ReadDouble();
     }
   }
 }
